requirements.txt
# Core
Flask==3.0.3
gunicorn==23.0.0

# LangChain stack (متوافق مع google-genai 0.4.1)
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.1
langsmith==0.1.71

# Google GenAI
google-generativeai==0.4.1
langchain-google-genai==0.0.11

# DB & Vectors
mysql-connector-python==9.0.0
chromadb==0.4.24

# NLP / Embeddings
sentence-transformers==2.2.2

# Utils
packaging==23.2          # <- هنا التغيير
requests==2.32.3
loguru==0.7.2
APScheduler==3.10.4
python-dotenv==1.0.1

# Scientific (Python 3.11)
numpy==1.26.4
pandas==2.3.1

